D2GC: Cleaning up example + README.md
I cleaned up the D2GC Example a bit to get rid of some extra
printed information that's either redundant or not really needed
for the purposes of the example. If people want the details, they
can run the perf_test.

Adding an initial README.md for the example/graph directory that
describes the D2GC example.
